chore: cleanup part 1
