--- neovide_backtraces.log
@@ -1,60 +1,0 @@
-2023-07-15 11:25:58 - Neovide panicked with the message 'Could not parse event from neovim: invalid event format for event 'grid_cursor_goto' - [Integer(PosInt(3)), Integer(NegInt(-1)), Integer(PosInt(20))] - invalid u64 format -1'. (File: src/error_handling.rs; Line: 5, Column: 5)
-   0: <unknown>
-   1: <unknown>
-   2: <unknown>
-   3: <unknown>
-   4: <unknown>
-   5: <unknown>
-   6: <unknown>
-   7: <unknown>
-   8: <unknown>
-   9: <unknown>
-  10: <unknown>
-  11: <unknown>
-  12: <unknown>
-  13: <unknown>
-  14: <unknown>
-  15: <unknown>
-  16: start_thread
-  17: __clone3
-
-2023-07-15 18:25:18 - Neovide panicked with the message 'Could not parse event from neovim: invalid event format for event 'grid_cursor_goto' - [Integer(PosInt(3)), Integer(NegInt(-1)), Integer(PosInt(20))] - invalid u64 format -1'. (File: src/error_handling.rs; Line: 5, Column: 5)
-   0: <unknown>
-   1: <unknown>
-   2: <unknown>
-   3: <unknown>
-   4: <unknown>
-   5: <unknown>
-   6: <unknown>
-   7: <unknown>
-   8: <unknown>
-   9: <unknown>
-  10: <unknown>
-  11: <unknown>
-  12: <unknown>
-  13: <unknown>
-  14: <unknown>
-  15: <unknown>
-  16: start_thread
-  17: __clone3
-
-2023-07-15 18:36:00 - Neovide panicked with the message 'Could not parse event from neovim: invalid event format for event 'grid_cursor_goto' - [Integer(PosInt(3)), Integer(NegInt(-1)), Integer(PosInt(16))] - invalid u64 format -1'. (File: src/error_handling.rs; Line: 5, Column: 5)
-   0: <unknown>
-   1: <unknown>
-   2: <unknown>
-   3: <unknown>
-   4: <unknown>
-   5: <unknown>
-   6: <unknown>
-   7: <unknown>
-   8: <unknown>
-   9: <unknown>
-  10: <unknown>
-  11: <unknown>
-  12: <unknown>
-  13: <unknown>
-  14: <unknown>
-  15: <unknown>
-  16: start_thread
-  17: __clone3
-
