+++ src/lib.rs
@@ -4,3 +4,4 @@
 pub mod assets;
 pub mod auth;
 pub mod launcher;
+pub mod parser;
